STOR-46 rename packages
"custom_converter" to "customConverter"
"file_upload" to "fileUpload"
