+++ src/main/java/com/butbetter/storage/StorageApplication.java
@@ -1,6 +1,6 @@
 package com.butbetter.storage;
 
-import com.butbetter.storage.file_upload.properties.StorageProperties;
+import com.butbetter.storage.fileUpload.properties.StorageProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
