+++ src/main/java/com/butbetter/storage/customConverter/BeanOffsetDateTimeConverter.java
@@ -1,7 +1,6 @@
-package com.butbetter.storage.custom_converter;
+package com.butbetter.storage.customConverter;
 
 import com.opencsv.bean.AbstractBeanField;
-import com.opencsv.exceptions.CsvConstraintViolationException;
 import com.opencsv.exceptions.CsvDataTypeMismatchException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
