+++ src/main/java/com/butbetter/storage/fileUpload/FileStorageService.java
@@ -1,10 +1,10 @@
-package com.butbetter.storage.file_upload;
+package com.butbetter.storage.fileUpload;
 
 import com.butbetter.storage.csv.CSVImportService;
 import com.butbetter.storage.csv.exceptions.FaultyCSVException;
-import com.butbetter.storage.file_upload.exceptions.StorageException;
-import com.butbetter.storage.file_upload.exceptions.StorageFileNotFoundException;
-import com.butbetter.storage.file_upload.properties.StorageProperties;
+import com.butbetter.storage.fileUpload.exceptions.StorageException;
+import com.butbetter.storage.fileUpload.exceptions.StorageFileNotFoundException;
+import com.butbetter.storage.fileUpload.properties.StorageProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
